+++ R/qc.R
@@ -1,14 +1,15 @@
 #' Quality Check
 #'
-#' Run all quality checks on an Excel STAR template.
+#' Run all quality checks on one or more Excel STAR templates.
 #'
-#' @param file filename of an Excel STAR template.
+#' @param x filename of an Excel STAR template or a directory containing Excel
+#'        STAR templates.
 #' @param stop whether to stop if test fails.
 #' @param quiet whether to suppress messages.
 #'
 #' @return
-#' \code{TRUE} if all tests succeed, otherwise an error message (if
-#' \code{stop = TRUE}) or FALSE and a warning message (if \code{stop = FALSE}).
+#' Logical \code{TRUE} or \code{FALSE} indicating whether all tests succeeded,
+#' or a logical vector if \code{x} is a directory.
 #'
 #' @seealso
 #' The checks are run in the following order:
@@ -25,20 +26,39 @@
 #'
 #' @examples
 #' \dontrun{
+#'
+#' # Filename
 #' qc("STAR_2019_HKE_5.xlsx")
+#'
+#' # Directory
+#' ok <- qc("working_group")
+#' ok[!ok]
 #' }
 #'
 #' @export
 
-qc <- function(file, stop=TRUE, quiet=FALSE)
+qc <- function(x, stop=TRUE, quiet=FALSE)
 {
-  ## Start with success TRUE and later flip it to FALSE if any test fails
-  s <- TRUE
+  if(dir.exists(x))
+  {
+    files <- dir(x, pattern="\\.xlsx?$", full.names=TRUE)
+    s <- suppressWarnings(sapply(files, qc, stop=FALSE, quiet=FALSE))
+    names(s) <- basename(names(s))
+  }
+  else if(file.exists(x))
+  {
+    ## Start with success TRUE and later flip it to FALSE if any test fails
+    s <- TRUE
 
-  s <- s && qc.exists(file, stop=stop, quiet=quiet)
-  s <- s && qc.xlsx(file, stop=stop, quiet=quiet)
-  s <- s && qc.star(file, stop=stop, quiet=quiet)
-  s <- s && qc.vpa(file, stop=stop, quiet=quiet)
+    s <- s && qc.exists(x, stop=stop, quiet=quiet)
+    s <- s && qc.xlsx(x, stop=stop, quiet=quiet)
+    s <- s && qc.star(x, stop=stop, quiet=quiet)
+    s <- s && qc.vpa(x, stop=stop, quiet=quiet)
+  }
+  else
+  {
+    stop("'x' must be an existing filename or directory")
+  }
 
-  s
+  invisible(s)
 }